requirements.txt
Flask
flask-cors
deep-translator
SpeechRecognition
python-dotenv